requirements.txt
requests~=2.27.1
pibooth~=2.0.5
Pillow~=9.1.0